requirements.txt
picamera
python-telegram-bot
